# openapi.json
{
  "components": {
    "schemas": {
...
      "PortalAuthResponse": {
        "properties": {
          "expires_at": {
            "type": "string"
          },
          "organization_id": {
            "type": "integer"
          },
          "patient_id": {
            "type": "integer"
          },
          "role": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token",
          "role",
          "patient_id",
          "organization_id",
          "expires_at"
        ],
        "type": "object"
      },
...
      "PortalLoginRequest": {
        "properties": {
          "email": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "in": "cookie",
        "name": "sessionid",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "CareOS API",
    "version": "0.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/portal/auth/login/": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortalLoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PortalAuthResponse"
                }
              }
            },
            "description": "OK"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          }
        },
        "summary": "Portal login"
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:8000"
    }
  ]
}